lib: extract Bpipe code from of compat.cc
this moved all of the windows specific Bpipe code into its own
translation unit, separating it from the windows compatibility
layer.
This also updates OpenBpipe() to use the new signature, however we do
not yet support passing custom environment variables to windows child
processes.
